Move javascript to `app/assets` folder
Another step in preparation for the migration away from Webpack(er). The
new Rails default is for JS to live in `app/javascript`, but given that
all our other frontend assets are in `app/assets`, it makes sense to
have them there.
