+++ app/assets/javascript/components/filters_component/filters_component.js
