+++ app/assets/javascript/components/map_component/cluster.test.js
