+++ app/assets/javascript/components/map_component/map.js
