+++ app/assets/javascript/components/map_component/map.test.js
