+++ app/assets/javascript/components/map_component/map_component.js
