+++ app/assets/javascript/components/map_component/marker/__mocks__/service.js
