+++ app/assets/javascript/components/map_component/marker/service.js
