+++ app/assets/javascript/components/map_component/marker/template.js
