+++ app/assets/javascript/components/map_component/sidebar.js
