+++ app/assets/javascript/components/map_component/sidebar.test.js
