+++ app/assets/javascript/components/searchable_collection_component/searchable_collection_component.test.js
