+++ app/assets/javascript/js_components/__mocks__/loader.js
