+++ app/assets/javascript/js_components/clipboard/clipboard.js
