+++ app/assets/javascript/js_components/loadingIndicator/loadingIndicator.js
