+++ app/assets/javascript/js_components/loadingIndicator/loadingIndicator.test.js
