+++ app/assets/javascript/js_components/locationFinder/locationFinder.js
