+++ app/assets/javascript/js_components/locationFinder/locationFinder.test.js
