+++ app/assets/javascript/js_components/manageQualifications/manageQualifications.js
