+++ app/assets/javascript/js_components/panel/panel.js
