+++ app/assets/javascript/js_components/showHiddenContent/showHiddenContent.js
