+++ app/assets/javascript/js_components/showHiddenContent/showHiddenContent.test.js
