+++ app/assets/javascript/js_components/trackedLink/trackedLink.js
