+++ app/assets/javascript/js_components/uploadDocuments/uploadDocuments.js
