+++ app/assets/javascript/js_components/uploadDocuments/uploadDocuments.test.js
