+++ app/assets/javascript/js_components/utils.js
