+++ app/assets/javascript/lib/__mocks__/api.js
