+++ app/assets/javascript/lib/__mocks__/utils.js
