+++ app/assets/javascript/lib/events.js
