+++ app/assets/javascript/lib/logging.js
