+++ app/assets/javascript/lib/polyfill/closest.polyfill.js
