+++ app/assets/javascript/lib/polyfill/from.polyfill.js
