+++ app/assets/javascript/lib/utils.js
