+++ app/assets/javascript/lib/utils.test.js
