+++ config/webpacker.yml
@@ -1,7 +1,7 @@
 # Note: You must restart bin/webpack-dev-server for changes to take effect
 
 default: &default
-  source_path: app/frontend
+  source_path: app/assets/javascript
   source_entry_path: packs
   public_root_path: public
   public_output_path: packs
