+++ package.json
@@ -52,7 +52,7 @@
     "test": "concurrently \"yarn run js:test\" \"yarn run js:lint\" \"yarn run sass:lint\"",
     "js:test": "jest",
     "js:test:coverage": "jest --coverage",
-    "js:lint": "eslint ./app/frontend/src",
+    "js:lint": "eslint ./app/assets/javascript",
     "sass:lint": "yarn stylelint app/assets/stylesheets/**/*.scss -q",
     "backstop:execute": "backstop $CMD --config='app/frontend/backstop/backstop.config.js'",
     "visual:test:reference": "CMD=reference yarn run backstop:execute",
@@ -62,10 +62,10 @@
   },
   "jest": {
     "testMatch": [
-      "<rootDir>/app/frontend/src/**/*.test.js"
+      "<rootDir>/app/assets/javascript/**/*.test.js"
     ],
     "coveragePathIgnorePatterns": [
-      "<rootDir>/app/frontend/src/lib/polyfill/*"
+      "<rootDir>/app/assets/javascript/lib/polyfill/*"
     ],
     "testEnvironment": "jest-environment-jsdom"
   },
